LayoutTests/ietestcenter/Javascript/15.2.3.14-2-4-expected.txt

15.2.3.14-2-4 (TestCases/chapter15/15.2/15.2.3/15.2.3.14/15.2.3.14-2-4.js) Object.keys returns the standard built-in Array that is extensible

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL ES5Harness.preconditionPassed should be true (of type boolean). Was undefined (of type undefined).
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
